extract xtick
